requirements.txt
langchain-openai==0.3.23
langgraph==0.4.8
pre-commit==4.2.0
pymupdf4llm==0.0.25
langchain-community==0.3.25
unstructured==0.17.2
markdown==3.8
langchain-huggingface==0.3.0
sentence-transformers==4.1.0
chromadb==1.0.12
python-dotenv==1.1.0